+++ packages/nodes-base/nodes/Aws/IAM/test/ProcessGroupsResponse.test.ts
@@ -36,7 +36,7 @@ describe('processGroupsResponse', () => {
 		const result = await processGroupsResponse.call(mockContext, [], response);
 
 		expect(result).toHaveLength(2);
-		expect(result[0]).toEqual({ json: { GroupName: 'Group1', GroupId: '1' } });
-		expect(result[1]).toEqual({ json: { GroupName: 'Group2', GroupId: '2' } });
+		expect(result[0]).toEqual({ GroupName: 'Group1', GroupId: '1' });
+		expect(result[1]).toEqual({ GroupName: 'Group2', GroupId: '2' });
 	});
 });
